remove HTTPS requirement
This requirement was originally introduced due to dependency on
Reporting API (back when we had a reporting mode) which required
HTTPS. This dependency is no longer there.

Futher, now that we're merging permission delegation, which will
disable a number of features by default in embedded contexts.. We
do need a mechanims for non HTTPS pages to delegate permission.
That said, note that 'delegate' may not be sufficient in itself
as some features may have additional requirements in place.

E.g. in order for the embedee to gain access to Geolocation API
the embedder must delegate permission AND the embeddee must be
a 'potentially trustworthy' (HTTPS) origin.