Update header and allow attribute syntax
This introduces a new, heavily CSP-inspired syntax for the HTTP header
and the <iframe allow> attribute. It allows more complex container
policies to be specified in frames, without all of the verbosity and
escaping that JSON would require. The header format is changed to match
this, for consistency.

Examples are also updated; the parser section still needs to be
rewritten to take this into account; two issues are marked for that.

Fixes #78