Replace remaining instances of JSON-list syntax
There were a few places in the spec and feature list where an allowlist
was specified as ["self"], ["*"] or []. While not technically incorrect,
as they were not referring to the serialization of policies, they were
misleading. This replaces those with the hopefully less-misleading terms
from the new serialization: 'self', '*', and 'none'.

Fixes: #83