Edit the implementation detail for `touch-action`
The original idea behind `pinch-zoom` was that even without `pan-y` the user should be still able to zoom out and then scroll the page by touch scrolling outside `<iframe>` bounds. However, this seems to still be an issue if the <iframe> is covering most of the window space event at small page scale factors.  So we should only enforce `pan-y` when it does not exit.