Break the circular definitions
This fixes the issue where both this spec and the suggested changes to
the HTML spec declared that a document "has a" feature policy, but
neither spec explicitly declared what a feature policy "is" in the
absence of a document.

It also cleans up some of the other sloppy usage of the term "top-level
document" and "child document" to refer to documents in browsing
contexts instead.