Add explainer for implementing sandbox through FP
This adds a document which explains how all of the features which are currently controlled through iframe sandboxing can be made into policy-controlled features, with the `sandbox` attribute being used to influence the constructed container policy for the frame. This also allows those features to be controlled outside of sandboxes as well.