Fix inherited policy origin (#229)
Fix inherited policy algorithm

This corrects an issue with the "Define an inherited policy for feature"
algorithms where the wrong origin was used to compare against the
container policy and the parent document's feature policy. The origin of
the document in the frame (or of the document potentially loaded into
that frame) should be used, rather than the parent's origin.

This also cleans up some of the linked terms in the various algorithms
involved in this.