Fix inherited policy calculation (#234)
The inherited policy calculation previously would not properly disable
features in frames if they were disabled in the parent. This change
closes the hole where a top-level page could be delivered with a header
which disabled a feature, but then re-enable it in any frame. It also fixes
an edge case where specifying 'allow="feature"' could *disable* a feature
in a frame in which it would otherwise be enabled.

Fixes: #233