Default allowlists only effect child browsing contexts
This change updates the definition of a "default allowlist" to clarify
that it only effects feature availability in child browsing contexts.
That is, a feature is always available by default in a top-level
browsing context but may be specified with a default allowlist of '*' or
'self' in order to control whether the feature is available by default
in cross-origin documents in child browsing contexts.