Fix `Define an inherited policy for feature in container at origin`
The text implies only the inherited policy of the document should
be checked, but even if the containing document has the feature
enabled the document could set a permissions policy that denied access
to the specific origin. We must check the actual policy for the
containing document to know. Chrome is already doing this.